# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EAMS API",
    "version": "1.0.0",
    "description": "Event Access Management System API Documentation",
    "contact": {
      "name": "API Support",
      "email": "support@eams.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Event": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "startDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" },
          "venue": {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "address": { "type": "string" },
              "city": { "type": "string" }
            }
          },
          "status": { "type": "string", "enum": ["draft", "published", "ongoing", "completed", "cancelled"] },
          "settings": {
            "type": "object",
            "properties": {
              "currency": { "type": "string" },
              "requirePhotoVerification": { "type": "boolean" }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "error": { "type": "string" }
        }
      }
    }
  },
  "security": [{ "BearerAuth": [] }],
...
  "paths": {
...
    "/events": {
      "get": {
        "tags": ["Events"],
        "summary": "Get all events",
        "parameters": [
          { "in": "query", "name": "status", "schema": { "type": "string" } },
          { "in": "query", "name": "page", "schema": { "type": "integer" } },
          { "in": "query", "name": "limit", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "events": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } },
                        "total": { "type": "integer" },
                        "pages": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Events"],
        "summary": "Create a new event (Organiser/Admin)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "startDate", "venue"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "startDate": { "type": "string", "format": "date-time" },
                  "endDate": { "type": "string", "format": "date-time" },
                  "venue": { "type": "object" },
                  "settings": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Event created" },
          "400": { "$ref": "#/components/schemas/Error" }
        }
      }
    },
...
  }
}